+++ .gitignore
@@ -2,3 +2,4 @@ LFW
 __pycache__
 pg-data
 pgadmin-data
+pgadmin4-data
