--- routes/__init__.py
@@ -1,6 +1,0 @@
-from flask import Blueprint
-routes = Blueprint('routes', __name__)
-
-from .face import *
-from .partner import *
-from .user import *
